v0.0.4 调整结构添加了create和finding操作 #4 rt.
